fix warnings related to frameTime & frameNumber
long -> uint32_t frameNumber;
long -> uint16_t frameTime;

+++ src/Avara.cpp
@@ -76,7 +76,7 @@ int main(int argc, char *argv[]) {
         } else if (arg == "-h" || arg == "--host") {
             host = true;
         } else if (arg == "-f" || arg == "--frametime") {
-            long frameTime = atol(argv[++i]);  // pre-inc to next arg
+            uint16_t frameTime = atol(argv[++i]);  // pre-inc to next arg
             app->GetGame()->SetFrameTime(frameTime);
         } else if (arg == "--command") {
             textCommand = std::string(argv[++i]);
@@ -100,7 +100,7 @@ int main(int argc, char *argv[]) {
     } else if(connectAddress.size() > 0) {
         app->GetNet()->ChangeNet(kClientNet, connectAddress);
     }
-    
+
     // the mainloop should be sufficiently fast so that frames occur near their scheduled time
     mainloop(app->GetGame()->frameTime / 4);
